separate history expansion
