+++ src/history_expand.h
@@ -1,0 +1,6 @@
+#ifndef HISTORY_EXPAND_H
+#define HISTORY_EXPAND_H
+
+char *expand_history(const char *line);
+
+#endif /* HISTORY_EXPAND_H */
