+++ src/startup.c
@@ -6,6 +6,7 @@
 #include "startup.h"
 #include "parser.h"
 #include "history.h"
+#include "history_expand.h"
 #include "var_expand.h"
 #include "execute.h"
 #include "util.h"
