+++ src/var_expand.h
@@ -5,7 +5,6 @@ char *expand_var(const char *token);
 char *ansi_unescape(const char *src);
 char *expand_simple(const char *token);
 char **split_fields(const char *text, int *count);
-char *expand_history(const char *line);
 char **expand_braces(const char *word, int *count_out);
 char *expand_prompt(const char *prompt);
